Conflicts:
	app/views/users/invitations/new.html.erb

+++ lib/tasks/import_transform.rake
@@ -7,6 +7,11 @@ namespace :db do
     datafile =  args[:datafile]
     debug_output = args[:debug_output]
 
+    if !File.file?(datafile)
+      abort( "Aborting import : #{datafile} could not be loaded, check the filename!")
+    end
+
+
     # - IMPORTANT: SEED DATA ONLY
     puts 'Ensure all tables are truncated'
     ActiveRecord::Base.connection.execute("TRUNCATE TABLE tokens RESTART IDENTITY;")
@@ -54,8 +59,11 @@ namespace :db do
     statements.pop  # the last empty statement
     puts "Executing #{statements.size} statements" unless debug_output==true
     ActiveRecord::Base.transaction do
-      statements.each do |statement|
+        count=0
+        statements.each do |statement|
+        count+=1
         puts "Running: #{statement}\n" if debug_output==true
+        puts "Running statement: #{count}" unless debug_output==true
         rc.exec(statement)
 
         if statement =~ /STDIN/
